feat(web): remove org scope from host settings slice
