+++ apps/web/app/(dashboard)/hosts/[id]/host-terminal-launcher.tsx
@@ -10,13 +10,12 @@ import { useSession } from '@/lib/auth/client'
 import type { HostWithAgent } from '@/lib/actions/agents-core'
 
 interface Props {
-  scopeId: string
   host: HostWithAgent
   directAccess: boolean
   accessDeniedReason: string | null
 }
 
-export function HostTerminalLauncher({ scopeId: _orgId, host, directAccess, accessDeniedReason }: Props) {
+export function HostTerminalLauncher({ host, directAccess, accessDeniedReason }: Props) {
   const { openTerminal } = useTerminalPanel()
   const { data: session } = useSession()
