+++ apps/web/components/notes/note-editor-dialog.tsx
@@ -33,7 +33,6 @@ import { NOTE_CATEGORIES, type NoteCategory } from '@/lib/db/schema'
 
 interface NewNoteProps {
   mode: 'create'
-  scopeId: string
   hostId: string
   open: boolean
   onOpenChange: (open: boolean) => void
@@ -41,7 +40,6 @@ interface NewNoteProps {
 
 interface EditNoteProps {
   mode: 'edit'
-  scopeId: string
   hostId: string
   noteId: string
   initial: {
@@ -67,7 +65,7 @@ const CATEGORY_LABELS: Record<NoteCategory, string> = {
 }
 
 export function NoteEditorDialog(props: Props) {
-  const { mode, scopeId, hostId, open, onOpenChange } = props
+  const { mode, hostId, open, onOpenChange } = props
   const queryClient = useQueryClient()
 
   const initial = useMemo(
@@ -98,12 +96,12 @@ export function NoteEditorDialog(props: Props) {
   }
 
   const invalidateHostNotes = () => {
-    queryClient.invalidateQueries({ queryKey: ['notes-for-host', scopeId, hostId] })
+    queryClient.invalidateQueries({ queryKey: ['notes-for-host', hostId] })
   }
 
   const createMutation = useMutation({
     mutationFn: async () => {
-      const result = await createNote(scopeId, {
+      const result = await createNote({
         title: title.trim(),
         body,
         category,
@@ -123,7 +121,7 @@ export function NoteEditorDialog(props: Props) {
   const updateMutation = useMutation({
     mutationFn: async () => {
       if (mode !== 'edit') throw new Error('wrong mode')
-      const result = await updateNote(scopeId, props.noteId, {
+      const result = await updateNote(props.noteId, {
         title: title.trim(),
         body,
         category,
@@ -133,7 +131,7 @@ export function NoteEditorDialog(props: Props) {
       // Privacy is author-only; skip the call if the flag didn't change or the
       // user isn't the author (admins don't see the toggle at all).
       if (props.initial.isAuthor && props.initial.isPrivate !== isPrivate) {
-        const privacyResult = await toggleNotePrivate(scopeId, props.noteId, isPrivate)
+        const privacyResult = await toggleNotePrivate(props.noteId, isPrivate)
         if ('error' in privacyResult) throw new Error(privacyResult.error)
       }
       return result.note
@@ -141,7 +139,7 @@ export function NoteEditorDialog(props: Props) {
     onSuccess: () => {
       invalidateHostNotes()
       if (mode === 'edit') {
-        queryClient.invalidateQueries({ queryKey: ['note-revisions', scopeId, props.noteId] })
+        queryClient.invalidateQueries({ queryKey: ['note-revisions', props.noteId] })
       }
       onOpenChange(false)
     },
@@ -230,7 +228,7 @@ export function NoteEditorDialog(props: Props) {
                 <p className="text-xs text-muted-foreground">
                   {isPrivate
                     ? 'Only you and super admins can read this note.'
-                    : 'Everyone in your org who can see this host.'}
+                    : 'Everyone in this CT-Ops instance who can see this host.'}
                 </p>
               </div>
             )}
