+++ apps/web/components/notes/pinned-notes-card.tsx
@@ -8,7 +8,6 @@ import { listNotesForHost } from '@/lib/actions/notes'
 import { NoteCard } from './note-card'
 
 interface Props {
-  scopeId: string
   hostId: string
   currentUserId: string
   userRole: string
@@ -21,15 +20,14 @@ interface Props {
 // the Overview doesn't carry an empty section. Cards render in compact mode
 // (collapsed body) so a pin'd runbook doesn't push metrics off the fold.
 export function PinnedNotesCard({
-  scopeId,
   hostId,
   currentUserId,
   userRole,
   onViewAll,
 }: Props) {
   const { data: notes = [] } = useQuery({
-    queryKey: ['notes-for-host', scopeId, hostId],
-    queryFn: () => listNotesForHost(scopeId, hostId),
+    queryKey: ['notes-for-host', hostId],
+    queryFn: () => listNotesForHost(hostId),
   })
 
   const pinned = useMemo(() => notes.filter((n) => n.isPinned), [notes])
@@ -61,7 +59,6 @@ export function PinnedNotesCard({
           <NoteCard
             key={note.id}
             note={note}
-            scopeId={scopeId}
             hostId={hostId}
             currentUserId={currentUserId}
             userRole={userRole}
