+++ apps/web/lib/actions/host-groups.ts
@@ -1,246 +1,90 @@
 'use server'
 
-import { logError } from '@/lib/logging'
-import { requireOrgAccess, requireOrgWriteAccess } from '@/lib/actions/action-auth'
+import { getRequiredSession } from '@/lib/auth/session'
+import { resolveCurrentActionScope } from './action-scope'
+import {
+  addHostToGroup as addHostToGroupCore,
+  createGroup as createGroupCore,
+  deleteGroup as deleteGroupCore,
+  getGroup as getGroupCore,
+  listGroups as listGroupsCore,
+  listGroupsForHost as listGroupsForHostCore,
+  listHostsInGroup as listHostsInGroupCore,
+  removeHostFromGroup as removeHostFromGroupCore,
+  updateGroup as updateGroupCore,
+  type HostGroupWithCount,
+  type HostGroupWithMembers,
+} from './host-groups-core'
 
-import { z } from 'zod'
-import { db } from '@/lib/db'
-import { hostGroups, hostGroupMembers, hosts } from '@/lib/db/schema'
-import { eq, and, isNull, sql } from 'drizzle-orm'
-import type { HostGroup } from '@/lib/db/schema'
-import type { Host } from '@/lib/db/schema'
-
-export type HostGroupWithCount = HostGroup & { hostCount: number }
-export type HostGroupWithMembers = HostGroup & { members: Host[] }
-
-const groupSchema = z.object({
-  name: z.string().min(1).max(255),
-  description: z.string().max(1000).optional(),
-})
-
-// ── Group CRUD ─────────────────────────────────────────────────────────────
+export type { HostGroupWithCount, HostGroupWithMembers } from './host-groups-core'
 
 export async function createGroup(
-  orgId: string,
+  scopeId: string,
   input: unknown,
-): Promise<{ success: true; group: HostGroup } | { error: string }> {
-  await requireOrgWriteAccess(orgId)
-  const parsed = groupSchema.safeParse(input)
-  if (!parsed.success) {
-    return { error: parsed.error.issues[0]?.message ?? 'Invalid input' }
-  }
-  try {
-    const rows = await db
-      .insert(hostGroups)
-      .values({
-        organisationId: orgId,
-        name: parsed.data.name,
-        description: parsed.data.description ?? null,
-      })
-      .returning()
-    const group = rows[0]
-    if (!group) return { error: 'Failed to create group' }
-    return { success: true, group }
-  } catch (err) {
-    logError('Failed to create group:', err)
-    return { error: 'Failed to create group' }
-  }
+): Promise<{ success: true; group: import('@/lib/db/schema').HostGroup } | { error: string }> {
+  return createGroupCore(scopeId, input)
 }
 
 export async function updateGroup(
-  orgId: string,
+  scopeId: string,
   groupId: string,
   input: unknown,
 ): Promise<{ success: true } | { error: string }> {
-  await requireOrgWriteAccess(orgId)
-  const parsed = groupSchema.safeParse(input)
-  if (!parsed.success) {
-    return { error: parsed.error.issues[0]?.message ?? 'Invalid input' }
-  }
-  try {
-    const result = await db
-      .update(hostGroups)
-      .set({ name: parsed.data.name, description: parsed.data.description ?? null, updatedAt: new Date() })
-      .where(and(eq(hostGroups.id, groupId), eq(hostGroups.organisationId, orgId), isNull(hostGroups.deletedAt)))
-      .returning({ id: hostGroups.id })
-
-    if (result.length === 0) return { error: 'Group not found' }
-    return { success: true }
-  } catch (err) {
-    logError('Failed to update group:', err)
-    return { error: 'Failed to update group' }
-  }
+  return updateGroupCore(scopeId, groupId, input)
 }
 
 export async function deleteGroup(
-  orgId: string,
+  scopeId: string,
   groupId: string,
 ): Promise<{ success: true } | { error: string }> {
-  await requireOrgWriteAccess(orgId)
-  try {
-    // Soft-delete the group
-    const result = await db
-      .update(hostGroups)
-      .set({ deletedAt: new Date(), updatedAt: new Date() })
-      .where(and(eq(hostGroups.id, groupId), eq(hostGroups.organisationId, orgId), isNull(hostGroups.deletedAt)))
-      .returning({ id: hostGroups.id })
-
-    if (result.length === 0) return { error: 'Group not found' }
-
-    // Soft-delete all memberships for this group
-    await db
-      .update(hostGroupMembers)
-      .set({ deletedAt: new Date(), updatedAt: new Date() })
-      .where(and(eq(hostGroupMembers.groupId, groupId), eq(hostGroupMembers.organisationId, orgId), isNull(hostGroupMembers.deletedAt)))
-
-    return { success: true }
-  } catch (err) {
-    logError('Failed to delete group:', err)
-    return { error: 'Failed to delete group' }
-  }
+  return deleteGroupCore(scopeId, groupId)
 }
 
-export async function listGroups(orgId: string): Promise<HostGroupWithCount[]> {
-  await requireOrgAccess(orgId)
-  const groups = await db.query.hostGroups.findMany({
-    where: and(eq(hostGroups.organisationId, orgId), isNull(hostGroups.deletedAt)),
-    orderBy: (g, { asc }) => [asc(g.name)],
-  })
-
-  // Get host counts for each group in one query
-  const counts = await db
-    .select({
-      groupId: hostGroupMembers.groupId,
-      count: sql<number>`cast(count(*) as int)`,
-    })
-    .from(hostGroupMembers)
-    .where(and(eq(hostGroupMembers.organisationId, orgId), isNull(hostGroupMembers.deletedAt)))
-    .groupBy(hostGroupMembers.groupId)
-
-  const countMap = new Map(counts.map((c) => [c.groupId, c.count]))
-
-  return groups.map((g) => ({ ...g, hostCount: countMap.get(g.id) ?? 0 }))
+export async function listGroups(
+  ...args: [] | [string]
+): Promise<HostGroupWithCount[]> {
+  const session = await getRequiredSession()
+  const currentScope = args[0] ?? resolveCurrentActionScope(session)
+  return listGroupsCore(currentScope)
 }
 
-export async function getGroup(orgId: string, groupId: string): Promise<HostGroupWithMembers | null> {
-  await requireOrgAccess(orgId)
-  const group = await db.query.hostGroups.findFirst({
-    where: and(eq(hostGroups.id, groupId), eq(hostGroups.organisationId, orgId), isNull(hostGroups.deletedAt)),
-  })
-  if (!group) return null
-
-  const members = await listHostsInGroup(orgId, groupId)
-  return { ...group, members }
+export async function getGroup(
+  scopeId: string,
+  groupId: string,
+): Promise<HostGroupWithMembers | null> {
+  return getGroupCore(scopeId, groupId)
 }
-
-// ── Membership ────────────────────────────────────────────────────────────
 
 export async function addHostToGroup(
-  orgId: string,
-  groupId: string,
-  hostId: string,
+  ...args: [string, string] | [string, string, string]
 ): Promise<{ success: true } | { error: string }> {
-  await requireOrgWriteAccess(orgId)
-  try {
-    // Check if already a member (including soft-deleted — restore if so)
-    const existing = await db.query.hostGroupMembers.findFirst({
-      where: and(
-        eq(hostGroupMembers.groupId, groupId),
-        eq(hostGroupMembers.hostId, hostId),
-        eq(hostGroupMembers.organisationId, orgId),
-      ),
-    })
-
-    if (existing) {
-      if (!existing.deletedAt) return { error: 'Host is already in this group' }
-      // Restore soft-deleted membership
-      await db
-        .update(hostGroupMembers)
-        .set({ deletedAt: null, updatedAt: new Date() })
-        .where(eq(hostGroupMembers.id, existing.id))
-      return { success: true }
-    }
-
-    await db.insert(hostGroupMembers).values({
-      organisationId: orgId,
-      groupId,
-      hostId,
-    })
-    return { success: true }
-  } catch (err) {
-    logError('Failed to add host to group:', err)
-    return { error: 'Failed to add host to group' }
-  }
+  const session = await getRequiredSession()
+  const [currentScope, groupId, hostId] =
+    args.length === 3 ? args : [resolveCurrentActionScope(session), args[0], args[1]]
+  return addHostToGroupCore(currentScope, groupId, hostId)
 }
 
 export async function removeHostFromGroup(
-  orgId: string,
-  groupId: string,
-  hostId: string,
+  ...args: [string, string] | [string, string, string]
 ): Promise<{ success: true } | { error: string }> {
-  await requireOrgWriteAccess(orgId)
-  try {
-    const result = await db
-      .update(hostGroupMembers)
-      .set({ deletedAt: new Date(), updatedAt: new Date() })
-      .where(
-        and(
-          eq(hostGroupMembers.groupId, groupId),
-          eq(hostGroupMembers.hostId, hostId),
-          eq(hostGroupMembers.organisationId, orgId),
-          isNull(hostGroupMembers.deletedAt),
-        ),
-      )
-      .returning({ id: hostGroupMembers.id })
-
-    if (result.length === 0) return { error: 'Membership not found' }
-    return { success: true }
-  } catch (err) {
-    logError('Failed to remove host from group:', err)
-    return { error: 'Failed to remove host from group' }
-  }
+  const session = await getRequiredSession()
+  const [currentScope, groupId, hostId] =
+    args.length === 3 ? args : [resolveCurrentActionScope(session), args[0], args[1]]
+  return removeHostFromGroupCore(currentScope, groupId, hostId)
 }
 
-export async function listHostsInGroup(orgId: string, groupId: string): Promise<Host[]> {
-  await requireOrgAccess(orgId)
-  const members = await db.query.hostGroupMembers.findMany({
-    where: and(
-      eq(hostGroupMembers.groupId, groupId),
-      eq(hostGroupMembers.organisationId, orgId),
-      isNull(hostGroupMembers.deletedAt),
-    ),
-    columns: { hostId: true },
-  })
-
-  if (members.length === 0) return []
-
-  const hostIds = members.map((m) => m.hostId)
-
-  const hostRows = await db.query.hosts.findMany({
-    where: and(eq(hosts.organisationId, orgId), isNull(hosts.deletedAt)),
-  })
-
-  return hostRows.filter((h) => hostIds.includes(h.id))
+export async function listHostsInGroup(
+  scopeId: string,
+  groupId: string,
+): Promise<import('@/lib/db/schema').Host[]> {
+  return listHostsInGroupCore(scopeId, groupId)
 }
 
-export async function listGroupsForHost(orgId: string, hostId: string): Promise<HostGroup[]> {
-  await requireOrgAccess(orgId)
-  const members = await db.query.hostGroupMembers.findMany({
-    where: and(
-      eq(hostGroupMembers.hostId, hostId),
-      eq(hostGroupMembers.organisationId, orgId),
-      isNull(hostGroupMembers.deletedAt),
-    ),
-    columns: { groupId: true },
-  })
-
-  if (members.length === 0) return []
-
-  const groupIds = members.map((m) => m.groupId)
-
-  const groupRows = await db.query.hostGroups.findMany({
-    where: and(eq(hostGroups.organisationId, orgId), isNull(hostGroups.deletedAt)),
-  })
-
-  return groupRows.filter((g) => groupIds.includes(g.id))
+export async function listGroupsForHost(
+  ...args: [string] | [string, string]
+): Promise<import('@/lib/db/schema').HostGroup[]> {
+  const session = await getRequiredSession()
+  const [currentScope, hostId] =
+    args.length === 2 ? args : [resolveCurrentActionScope(session), args[0]]
+  return listGroupsForHostCore(currentScope, hostId)
 }
